Move python implementation to _frozendictpy
