+++ test/test_frozendict_subclass.py
@@ -1,4 +1,4 @@
-from frozendict.core import frozendict as FrozendictPyClass
+from frozendict._frozendictpy import frozendict as FrozendictPyClass
 from .common import FrozendictCommonTest
 from .subclass_only import FrozendictSubclassOnlyTest
 import frozendict as cool
